Make theme context stateful
